requirements.txt
requests
selenium
tinydb
proxy_list
pyvirtualdisplay
Pillow